Reorganized code, split problems into seperate files. Need to split libraries off into seperate files so they're available to problems.
